Removing warnings; fx odd bug; comment bad code

+++ Opcodes/date.c
@@ -139,7 +139,7 @@ static int readf(CSOUND *csound, READF *p)
 
 static int readfi(CSOUND *csound, READF *p)
 {
-    if (p->fd<=0)
+    if ((int)p->fd<=0)          /* SHOULD THIS BE ==NULL ?? */
       if (UNLIKELY(readf_init(csound, p)!= OK))
         return csound->InitError(csound, Str("readi failed to initialise"));
     return readf(csound, p);
